修正所有 Clippy 警告以通過 CI/CD 品質檢查
- 實作 Default trait 而非使用混淆的 default() 方法
- 實作 FromStr trait 而非使用混淆的 from_str() 方法
- 使用 push() 取代 push_str() 處理單字元
- 合併重複的 if 區塊邏輯
- 重新命名 add() 方法為 add_validator() 避免與標準 trait 混淆
- 移除不必要的解引用操作
- 實作 Display trait 取代直接的 ToString 實作
- 移除不必要的類型轉換
- 移除不必要的借用操作
- 使用函數引用取代冗餘的閉包

+++ src/redis/operations/cache.rs
@@ -116,7 +116,7 @@ impl<P: RedisPool> CacheManager<P> {
             .arg(&prefixed_key)
             .arg(&serialized)
             .arg("EX")
-            .arg(ttl_secs as u64)
+            .arg(ttl_secs)
             .query_async::<()>(&mut conn)
             .await
         {
@@ -149,7 +149,7 @@ impl<P: RedisPool> CacheOperations for CacheManager<P> {
                 .arg(&prefixed_key)
                 .arg(&serialized)
                 .arg("EX")
-                .arg(ttl as u64)
+                .arg(ttl)
                 .query_async::<()>(&mut conn)
                 .await
             {
